requirements.txt
mysql-connector-python
gradio
python-dotenv
dspy-ai